[NCL-8801]: Fix configs (application.yaml-s)
Initially, the config value of protected-tags-accepted-patterns was of
type List<String>. However, this results in build fail in case one
provides '[]' as a value -- this is a known, already closed, issue in
quarkus. The solution is to use Optional<List<String>> in such cases.

Also, prod application.yaml was split into 2 parts: static
(=non-changing) config values are in application.yaml, and all runtime
config values are present in runtime config.