Use ProcessStageUtils to log start of alignment pod phase
This commit uses `ProcessStageUtils` to log the start and end of the
starting-alignment-pod phase instead of using `userLogger`.

This is done because this type of logging is done for log-event-duration
to register the time it took inside a phase, and it requires extra
mdc values (`process_stage_step`, and `process_stage_name`) to know when the phase starts and ends.