[NCLSUP-1296] Parse pom.xml project.parent if field missing
For the scenario where PME is disabled, we read from the pom.xml to
extract the project's groupId, artifactId, and version.

In case the project's groupId and/or version is missing, we read it from
the project parent's groupId and/or version instead.

This commit also changes the way we extract the data by directly parsing
the Maven's pom.xml as an XML file instead of relying on Maven's
maven-help-plugin since it might fail when the parent cannot be
downloaded easily.